Rename the project to cs_ranking
