+++ src/code_winner/settings.py
@@ -37,7 +37,7 @@ INSTALLED_APPS = [
     'django.contrib.sessions',
     'django.contrib.messages',
     'django.contrib.staticfiles',
-    'ranking'
+    'cs_ranking'
 ]
 
 MIDDLEWARE_CLASSES = [
