+++ src/cs_ranking/templates/ranking/battle_result.html
