+++ src/cs_ranking/templates/ranking/index.html
