+++ src/cs_ranking/urls.py
@@ -2,7 +2,7 @@ from django.conf.urls import url
 
 from . import views
 
-app_name = 'ranking'
+app_name = 'cs_ranking'
 urlpatterns = [
     url(r'^$', views.index, name='index'),
     url(r'^battle/([0-9]+)$', views.battle_result, name='result'),
